moved commitlog to persistence directory
