+++ storage/src/lib.rs
@@ -7,6 +7,7 @@
 use std::convert::{TryFrom, TryInto};
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
+use std::str::FromStr;
 
 use failure::Fail;
 use rocksdb::Cache;
@@ -15,6 +16,7 @@ use slog::{error, info, Logger};
 use strum::IntoEnumIterator;
 use strum_macros::EnumIter;
 
+pub use action_file_storage::ActionFileStorage;
 use crypto::{
     base58::FromBase58CheckError,
     hash::{BlockHash, ChainId, ContextHash, FromBytesError, HashType},
@@ -23,9 +25,9 @@ use tezos_api::environment::{
     get_empty_operation_list_list_hash, TezosEnvironmentConfiguration, TezosEnvironmentError,
 };
 use tezos_api::ffi::{ApplyBlockResponse, CommitGenesisResult, PatchContext};
+use tezos_messages::Head;
 use tezos_messages::p2p::binary_message::{BinaryMessage, MessageHash, MessageHashError};
 use tezos_messages::p2p::encoding::prelude::BlockHeader;
-use tezos_messages::Head;
 
 pub use crate::block_meta_storage::{BlockMetaStorage, BlockMetaStorageKV, BlockMetaStorageReader};
 pub use crate::block_storage::{
@@ -42,14 +44,12 @@ pub use crate::operations_meta_storage::{OperationsMetaStorage, OperationsMetaSt
 pub use crate::operations_storage::{
     OperationKey, OperationsStorage, OperationsStorageKV, OperationsStorageReader,
 };
+use crate::persistent::{CommitLogError, DBError, Decoder, Encoder, SchemaError};
+use crate::persistent::ActionRecordError;
 pub use crate::persistent::database::{Direction, IteratorMode};
 use crate::persistent::sequence::SequenceError;
-use crate::persistent::ActionRecordError;
-use crate::persistent::{CommitLogError, DBError, Decoder, Encoder, SchemaError};
 pub use crate::predecessor_storage::PredecessorStorage;
 pub use crate::system_storage::SystemStorage;
-pub use action_file_storage::ActionFileStorage;
-use std::str::FromStr;
 
 pub mod action_file;
 pub mod action_file_storage;
@@ -68,7 +68,6 @@ pub mod predecessor_storage;
 pub mod skip_list;
 pub mod storage_backend;
 pub mod system_storage;
-pub mod commit_log;
 
 /// Extension of block header with block hash
 #[derive(Clone, Serialize, Deserialize, Debug, PartialEq)]
@@ -522,17 +521,17 @@ impl FromStr for KeyValueStoreBackend {
 }
 
 pub mod tests_common {
+    use std::{env, fs};
     use std::path::{Path, PathBuf};
     use std::sync::Arc;
-    use std::{env, fs};
 
     use failure::Error;
 
     use crate::block_storage;
     use crate::chain_meta_storage::ChainMetaStorage;
     use crate::mempool_storage::MempoolStorage;
-    use crate::persistent::sequence::Sequences;
     use crate::persistent::*;
+    use crate::persistent::sequence::Sequences;
     use crate::skip_list::{DatabaseBackedSkipList, Lane, ListValue};
 
     use super::*;
